// src/enums.ts
export enum Genre {
  Action = 'Action',
  Adventure = 'Adventure',
  Animation = 'Animation',
  Comedy = 'Comedy',
  Crime = 'Crime',
  Documentary = 'Documentary',
  Drama = 'Drama',
  Family = 'Family',
  Fantasy = 'Fantasy',
  History = 'History',
  Horror = 'Horror',
  Music = 'Music',
  Mystery = 'Mystery',
  Romance = 'Romance',
  ScienceFiction = 'Sci-Fi',
  TVMovie = 'TV Movie',
  Thriller = 'Thriller',
  War = 'War',
  Western = 'Western',
  ActionAdventure = 'Action & Adventure',
}

export enum MediaType {
  Movie,
  TvShow,
}